Removed dependence on specific game implementation

+++ src/Components/SquarePrimitiveComponent.h
@@ -7,14 +7,14 @@
 struct SquarePrimitiveComponent {
   glm::vec2 position;
   int layer;
-  int width;
-  int height;
+  float width;
+  float height;
   SDL_Color color;
   bool isFixed;
 
   explicit SquarePrimitiveComponent(glm::vec2 position = glm::vec2(0, 0), int layer = 0,
-                           int width = 0, int height = 0,
-                           SDL_Color color = {255, 255, 255, 255},
+                                    float width = 0, float height = 0,
+                                    SDL_Color color = {255, 255, 255, 255},
                            bool isFixed = true)
       : position(position),
         layer(layer),
